moved the module files into a package, fixed imports
